Internet_shitshow
AUTH: HammerSoup
NAME: IN_SS/Internet_shitshow
DESC: causes a bunch of random traffic stuff to target ip/lots of random traffic
DATE: 2/11/2021

IN_SS/Internet_shitshow is a bash script that will randomly generate web traffic based on 3 different properties that are defined in the settings.
1. the amount of time before traffic is sent
2. the number of connections that are sent
3. the type of connection that is sent 
The purpose of the project was to simulate internet traffic against the 2-Year system's students to disrupt their traffic filters.
I had a lot of fun making this project and using it within the 2-day deadline that I set for myself to make it. 
In the end, it showed some pretty interesting results.