php took too long
AUTH: HammerSoup
NAME: php took too long
DESC: does the entire assignment where i have to setup damn (vulnerable web application up and running)
DATE: 1/29/2021

This script was a really fun mini project to work on that I made for an assignment where I had to install DVWA (damn vulnerable web application).
The reason I made this script in the first place was when I completed the assignment and I thought it would be cool to automate. The setup was a long process to do and that's why I named it (php took too long).