Nmap script
AUTH: HammerSoup
DESC: Nmap script
DATE: 11/4/2021

A script that I made on the side because I wanted to automate my assignment because I didn't feel like staying in class that day. the joy of automation.
But to break this down, it simply runs Nmap against some Ips and formats it into a file, done! 
This was one of my first attempts at automation and in the end, it was really useful.