HTML Registration
AUTH: HammerSoup
DESC: HTML registration page
DATE: 4/6/2022
FILES: registration.html, registration.php

For the HTML page, It's pretty straightforward forward it's a registration page that gets credentials from the user and passes it off to the registration.php file to be processed into a MySQL database. In the registration.php file, the credentials go through multiple Harding methods in order to keep them secure like hashing before they are stored.

that's it in a nutshell
Anyway, it was a fun assignment and I finally got to learn and get hands-on with multiple languages.